Move ID length, generator to package text
