+++ pkg/text/text_test.go
@@ -15,3 +15,14 @@ func TestValidate(t *testing.T) {
 
 	assert.NoError(t, (&Text{Author: "a", Note: "n", URL: "u", Title: "t"}).Validate())
 }
+
+func TestRandomID(t *testing.T) {
+	set := map[string]bool{}
+	for i := 0; i < 10; i++ {
+		id, err := RandomID()
+		assert.NoError(t, err)
+		assert.Len(t, id, 8)
+		assert.NotContains(t, set, id)
+		set[id] = true
+	}
+}
